+++ libs/deepagents/src/profiles/harness/index.ts
@@ -1,19 +1,18 @@
 export {
-  // Interfaces
   type GeneralPurposeSubagentConfig,
   type HarnessProfile,
   type HarnessProfileOptions,
+  isHarnessProfile,
+  resolveMiddleware,
+  REQUIRED_MIDDLEWARE_NAMES,
+} from "./types.js";
+
+export { createHarnessProfile, EMPTY_HARNESS_PROFILE } from "./create.js";
+
+export {
   type HarnessProfileConfigData,
-  // Factory and serialization
-  createHarnessProfile,
-  serializeProfile,
-  parseHarnessProfileConfig,
-  // Zod schemas (for advanced users who need custom validation pipelines)
   harnessProfileConfigSchema,
   generalPurposeSubagentConfigSchema,
-  // Constants
-  EMPTY_HARNESS_PROFILE,
-  REQUIRED_MIDDLEWARE_NAMES,
-  // Internal helpers re-exported for registry module
-  resolveMiddleware,
-} from "./types.js";
+  parseHarnessProfileConfig,
+  serializeProfile,
+} from "./serialization.js";
